Add neo-tree floating version
We really missed this feature, so we add a tree browser back in.

Our old dotfiles was using nvim-tree, but we decided to switch to
neo-tree because it feels more stable. It was too intrusive (having to
disable netrw), and does not play well with other plugins (e.g. open
nvim, then press Ctrl-P. It will throw a weird error, and Ctrl-P will
show an empty window and fail to list the files in the project).

neo-tree is better and so we used that instead.

Another candidate was CHADtree but we dislike its dependency on python.