Fix lsp-zero complaining sometimes that a server is already init
Sometimes, lsp-zero will complain with such an error:

    [lsp-zero] Some language servers have been configured before
    you called the function .extened_lspconfig().

    Solution: Go to the place where you use lspconfig for the first time
    Call the .extend_lspconfig() function before you setup the language server

This error is not always thrown, so may need to exit and re-launch nvim
several times before getting the error.

The offending code is inside the config function of lsp-inlayhints.nvim.
This gives a clue as to why the error only throws sometimes: both do not
have direct dependencies with each other, so the order of
lsp-inlayhints.nvim and lsp-zero is random each time neovim is inited.

Our config calls require('lspconfig').tsserver.setup(), which is
forbidden to be called before lsp-zero get to initialize with
extend_lspconfig().

We fix this problem by moving the inlay hint setup for tsserver to mason
config function, which only initialises after lsp-zero.
lsp-inlayhints.nvim is only for setting up the virtual text UI, so with
our changes it also makes the config cleaner.