Try fix again inlay hint colors sometimes not consistent
Last time in commit fda1e14f0e0a6679af26ca6dbd2b386f067e7b15, we also
faced the issue where the inlay hint will have different colors on
different startups. We thought that the issue was that we didn't set
'LspInlayHint' highlight. But after applying the 'LspInlayHint' fix in
commit fda1e14f0e0a6679af26ca6dbd2b386f067e7b15, the issue persists.

Turns out it was because the loading sequence of color schemes and other
plugins are not consistent. This is usually not a problem, but for inlay
hints, they "cache" the colors[1], so depending on whether the color
scheme managed to change the color before the inlay plugin loads or not,
we may not always see the correct colors.

This commit finally fix it by setting a priority for the color schemes
to be always higher than the rest of the plugins.

[1] - https://github.com/lvimuser/lsp-inlayhints.nvim/blob/d981f65c9ae0b6062176f0accb9c151daeda6f16/lua/lsp-inlayhints/config.lua#L1-L20