fix: switch telescope grep quick postfix to `<C-c>` (#66)
`<C-p>` is normally used to go to previous entry in the search list while
in insert mode. We use this quite frequently, so it is bad to override
the shortcut with quick postfix.

`<C-c>` on the other hand, is less used. `<C-c>` automatically quits the
search window, but we usually just use `<Esc><Esc>` to do that, instead of
`<C-c>`. So it is an acceptable override.